api13 prelim
need ar api

+++ Automaton/UI/Debug/Tabs/ToolsTab.cs
@@ -4,7 +4,7 @@ using FFXIVClientStructs.FFXIV.Client.Game;
 using FFXIVClientStructs.FFXIV.Client.Game.Object;
 using FFXIVClientStructs.FFXIV.Client.UI.Agent;
 using FFXIVClientStructs.FFXIV.Component.GUI;
-using ImGuiNET;
+using Dalamud.Bindings.ImGui;
 using Lumina.Excel.Sheets;
 
 namespace Automaton.UI.Debug.Tabs;
@@ -53,7 +53,7 @@ internal unsafe class ToolsTab : DebugTab
         if (ImGui.Button("hg"))
         {
             var player = (FFXIVClientStructs.FFXIV.Client.Game.Character.Character*)GameObjectManager.Instance()->Objects.IndexSorted[0].Value;
-            player->GetStatusManager()->SetStatus(20, 149, 5.0f, 0, 0xE0000000, true);
+            player->GetStatusManager()->SetStatus(20, 149, 5.0f, 0, (ulong)0xE0000000, true);
         }
     }
 }
